Various cleanup jobs, change role names
I can't believe I didn't kill myself with excessive use of format like
that

I don't care that this new way requires python 3.6+, it's so much better
and you're basically a barbarian if you're still on 3.5 at this point,
solely because of this.

Also: role names are now [<colorcode>] rather than (<colorcode>) as
before.